Move ORMException and CacheException up one namespace level
This is important to keep BC
