+++ lib/Doctrine/ORM/Cache/Exception/InvalidResultCacheDriver.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Cache\Exception;
 
+use Doctrine\ORM\Cache\CacheException;
+
 final class InvalidResultCacheDriver extends \LogicException implements CacheException
 {
     public static function create() : self
