+++ lib/Doctrine/ORM/Cache/Exception/NonCacheableEntity.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Cache\Exception;
 
+use Doctrine\ORM\Cache\CacheException;
 use function sprintf;
 
 class NonCacheableEntity extends \LogicException implements CacheException
