+++ lib/Doctrine/ORM/Cache/Exception/QueryCacheUsesNonPersistentCache.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Doctrine\ORM\Cache\Exception;
 
 use Doctrine\Common\Cache\Cache;
+use Doctrine\ORM\Cache\CacheException;
 use function get_class;
 
 final class QueryCacheUsesNonPersistentCache extends \LogicException implements CacheException
