+++ lib/Doctrine/ORM/Cache/LockException.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Cache;
 
-use Doctrine\ORM\Cache\Exception\CacheException;
-
 /**
  * Lock exception for cache.
  */
