+++ lib/Doctrine/ORM/EntityNotFoundException.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM;
 
-use Doctrine\ORM\Exception\ORMException;
 use function implode;
 
 /**
