+++ lib/Doctrine/ORM/Exception/ConfigurationException.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Exception;
 
+use Doctrine\ORM\ORMException;
+
 interface ConfigurationException extends ORMException
 {
 }
