+++ lib/Doctrine/ORM/Exception/NotSupported.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Exception;
 
+use Doctrine\ORM\ORMException;
+
 final class NotSupported extends \LogicException implements ORMException
 {
     public static function create() : self
