+++ lib/Doctrine/ORM/Exception/UnexpectedAssociationValue.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Exception;
 
-use Doctrine\ORM\Cache\Exception\CacheException;
+use Doctrine\ORM\Cache\CacheException;
 use function sprintf;
 
 final class UnexpectedAssociationValue extends \LogicException implements CacheException
