+++ lib/Doctrine/ORM/Internal/Hydration/HydrationException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Internal\Hydration;
 
-use Doctrine\ORM\Exception\ORMException;
+use Doctrine\ORM\ORMException;
 use function implode;
 use function sprintf;
