+++ lib/Doctrine/ORM/Mapping/ClassMetadata.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Mapping;
 
-use Doctrine\ORM\Cache\Exception\CacheException;
+use Doctrine\ORM\Cache\CacheException;
 use Doctrine\ORM\Cache\Exception\NonCacheableEntityAssociation;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Mapping\Factory\NamingStrategy;
