+++ lib/Doctrine/ORM/Mapping/Exception/UnknownGeneratorType.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Mapping\Exception;
 
-use Doctrine\ORM\Exception\ORMException;
+use Doctrine\ORM\ORMException;
 
 final class UnknownGeneratorType extends \LogicException implements ORMException
 {
