+++ lib/Doctrine/ORM/PessimisticLockException.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM;
 
-use Doctrine\ORM\Exception\ORMException;
-
 class PessimisticLockException extends \RuntimeException implements ORMException
 {
     /**
