+++ lib/Doctrine/ORM/Query/QueryException.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Query;
 
-use Doctrine\ORM\Exception\ORMException;
 use Doctrine\ORM\Mapping\AssociationMetadata;
+use Doctrine\ORM\ORMException;
 use Doctrine\ORM\Query\AST\PathExpression;
 use function sprintf;
