+++ lib/Doctrine/ORM/Tools/Exception/MissingColumnException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Tools\Exception;
 
-use Doctrine\ORM\Exception\ORMException;
+use Doctrine\ORM\ORMException;
 use function sprintf;
 
 final class MissingColumnException extends \LogicException implements ORMException
