+++ tests/Doctrine/Tests/ORM/Mapping/XmlMappingDriverTest.php
@@ -49,7 +49,7 @@ class XmlMappingDriverTest extends AbstractMappingDriverTest
     }
 
     /**
-     * @expectedException \Doctrine\ORM\Cache\Exception\CacheException
+     * @expectedException \Doctrine\ORM\Cache\CacheException
      * @expectedExceptionMessage Entity association field "Doctrine\Tests\ORM\Mapping\XMLSLC#foo" not configured as part of the second-level cache.
      */
     public function testFailingSecondLevelCacheAssociation() : void
